changes after review, 2 reducer to 1, bigger Dialog
