+++ backend/src/routing/controllers/api/Abitur.ts
@@ -55,7 +55,6 @@ export default class Abitur {
             const args: (string | number)[] = [];
             const setString = buildSetString(args);
             sql = 'UPDATE abiturpruefungen SET' + setString +' WHERE id = ?';
-            console.log(req.body)
             args.push(id);
             updateData(sql, args, defaultUpdateCallback(res));
         };
