+++ frontend/src/reducer/FifthExamReducer.ts
@@ -1,9 +1,32 @@
-interface FifthExamState {
+export interface FifthExamState {
+  allExams: Array<ExamInterface>,
   preFilledDataIn5PKFormEditedByAdmin: object;
+}
+
+export interface ExamInterface {
+  examId: number,
+  examType: string,
+  studentName: string,
+  partnerStudentName: string,
+  updatedPartnerStudentName: string,
+  referenzfach: string,
+  updatedReferenzfach: string,
+  bezugsfach: string,
+  updatedBezugsfach: string,
+  examiner: string,
+  updatedExaminer: string,
+  topicArea: string,
+  updatedTopicArea: string,
+  problemQuestion: string,
+  updatedProblemQuestion: string,
+  presentationForm: string,
+  updatedPresentationForm: string,
+  approved: boolean,
 }
 
 const initialState: FifthExamState = {
   preFilledDataIn5PKFormEditedByAdmin: {},
+  allExams: [],
 };
 
 const FifthExamReducer = (state = initialState, action) => {
@@ -12,6 +35,8 @@ const FifthExamReducer = (state = initialState, action) => {
       return {
         ...state, preFilledDataIn5PKFormEditedByAdmin: action.prefilledData,
       };
+    case 'LOAD_ALL_EXAMS':
+      return { ...state, allExams: action.payload };
     default:
       return state;
   }
